Added generateInDevMode setting (#100)
This pull request adds a new setting to generate the service worker in dev mode(which is disabled by default). - `generateInDevMode: true/false`

This saves me from having to rebuild the application whenever I make changes to my service worker file. 

What do you think of this addition? Or is this something that was left out by intention?

- Recompiling seemed to work without having to restart the application
- You obviously still have to manually unregister the service worker and then open a new tab to get the changes